requirements.txt
# apex==0.1 # install with conda: conda install -c conda-forge/label/cf202003 nvidia-apex, or install from the source https://github.com/NVIDIA/apex
h5py==2.10.0
matplotlib==3.1.2
numpy==1.17.3
pandas==0.25.3
Pillow==7.1.2
pytorch-ignite==0.4.2
scipy==1.3.1
tensorboard==2.0.1
tensorboardX==1.9
# tensorflow-gpu==2.0.0
torch==1.3.0
torchvision==0.4.1
trains==0.16.2
cv2
csv
xlrd